Incporate file encryption using OpenSSL
Now the password file is encrypted using openssl. The encrypted file is
read, decrypted, and then the decrypted values are stored in a buffer
that is used for searching. Then once finished, the buffer is zeroed.

To do:
Clean up and refactor code - this is a basic first draft
Create system for first use/adding passwords
When finding a site match, show the site name that the partial matched
too (i.e. if 'as' entered, show 'lastpass')
Revise searching algorithm, check for cases of multiple hits etc.
Create system for exporting passwords to text file (in case they want to
use a different service - include instructions at top of exported file
for encrypting/decrypting with gpg/openssl command line tools)